services/rag-query/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
qdrant-client==1.12.1
boto3==1.35.24
pydantic==2.9.2
psycopg2-binary==2.9.9
prometheus-client==0.19.0
prometheus-fastapi-instrumentator